Changed from double to float precision

+++ src/math/Epsilon.cpp
@@ -1,3 +1,3 @@
 #include <math/Epsilon.h>
 
-const double Epsilon::E = Epsilon::compute();
+const float Epsilon::E = Epsilon::compute();
